+++ src/components/Video/YouTubeDataInput.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useState } from 'react';
-import { getYoutubeData } from '@/app/actions/youtubeFetch';
+import { getYoutubeData } from '@/app/actions/fetchYouTube';
 import InputComponent from '@/components/Nexon/InputComponent';
 import { buttonVariants } from '@/components/ui/button';
 import { cn } from '@/src/util/utils';
